requirements.txt
# requirements.txt

# ── Interface ──────────────────────────────────────────────────────────
streamlit>=1.32.0

# ── Data ───────────────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# ── Visualisation ──────────────────────────────────────────────────────
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Azure Blob Storage ─────────────────────────────────────────────────
azure-storage-blob>=12.19.0

# ── LangChain + LangGraph + Mistral ───────────────────────────────────
langchain>=0.2.0
langchain-core>=0.2.0
langchain-mistralai>=0.1.0
langgraph>=0.1.0

# ── Météo & Statistiques ───────────────────────────────────────────────
requests>=2.31.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── Utilitaires ────────────────────────────────────────────────────────
python-dotenv>=1.0.0
